move config files and update imports
