+++ pyhdx/config.ini
@@ -1,4 +1,4 @@
 [cluster]
 ip = 127.0.0.1
 port = 52123
-
+n_workers = 10
